+++ server/standalone/handlers/event.go
@@ -2,17 +2,13 @@ package handlers
 
 import (
 	"context"
-	"errors"
 
-	"github.com/go-playground/log/v7"
 	"github.com/owkin/orchestrator/lib/asset"
 	"github.com/owkin/orchestrator/lib/common"
 	servercommon "github.com/owkin/orchestrator/server/common"
 	"github.com/owkin/orchestrator/server/common/logger"
 	"github.com/owkin/orchestrator/server/standalone/dbal"
 	"github.com/owkin/orchestrator/server/standalone/interceptors"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 )
 
 // EventServer is the gRPC facade to Model manipulation
@@ -73,24 +69,5 @@ func (s *EventServer) SubscribeToEvents(param *asset.SubscribeToEventsParam, str
 	}
 
 	d := dbal.New(ctx, nil, conn, channel)
-	err = d.SubscribeToEvents(param.StartEventId, stream)
-
-	if errors.Is(err, context.Canceled) {
-		log.WithError(err).Info("SubscribeToEvents interrupted: context canceled")
-		return nil
-	}
-
-	st := status.Convert(err)
-	switch st.Code() {
-	case codes.Canceled:
-		log.WithError(err).Info("SubscribeToEvents interrupted: gRPC operation canceled")
-		return nil
-	case codes.Unavailable:
-		if st.Message() == "transport is closing" {
-			log.WithError(err).Infof("SubscribeToEvents interrupted: %s", st.Message())
-			return nil
-		}
-	}
-
-	return err
+	return d.SubscribeToEvents(param.StartEventId, stream)
 }